refactor(utils): move truncate string and remove duplicated
