+++ recoco/apps/projects/models.py
@@ -1077,15 +1077,4 @@ class ProjectSearchAdapter(watson.SearchAdapter):
         return strip_accents(content)
 
 
-def truncate_string(s, max_length):
-    """Truncate given string to max_length"""
-    if len(s) <= max_length:
-        return s
-    sub = s[:max_length]
-    if s[max_length] != " ":
-        # we are truncating last word, rewind to its begining
-        sub = sub[: sub.rfind(" ")]
-    return f"{sub}…"
-
-
 # eof
